requirements.txt
Flask==2.1.0
nidaqmx==0.6.0
pyserial==3.5
pyvesc==1.0.1
pandas==1.3.4
